requirements.txt
aiogram>=3.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
gigachat>=0.1.0
